.
Improved UX of navbar widget.

Updated to the latest version of the soul system with new features, and better stability.

Updated to the latest version of backend_dashboard with vital features, and more stability.

Updated to the latest version of modernuser, with more stability.

Improved UI of the contact-us, hero, info-community, and coming-soon widget.

Made footer widget to use dynamic information about services.

Added the info-moto widget, to display information about the organization's moto.

Added frequently-asked-questions widget, to provide useful info to the user.

Added target-diseases widget, to provide information on diseases to the user.

Created the widget-settings-manager widget to make it easier for widgets to manage vital settings that concern them, in the Web Faculty.

Made the global.css important stylesheet automatically insert to every HTML page.

Improved typing.