.
Updated to the latest version of backend_dashboard with minor performance improvements on the frontend.

Updated to the latest version of modernuser, with better UI, and platform stability.

Made the contact-us widget more responsive.

Improved typing.

Updated to the latest version of the soul system, with better platform stability.

Added ability to manage information about team member from the admin dashboard.