.
Updated to the latest version of the faculty of finance, with better typing, and stability.

Updated to the latest version of the modern faculty of users, with better typing, bug fixes, and stability.

Moved /login/ to /app/login/, for better organization, and ease of access, especially with the mobile app.

Improved design of the appointment booking widget.

Added ability to pay for appointments.

Improved typing.

Updated to the latest version of the soul system, with marked stability improvements, and better typing.

Created faculty of health to manage aspects of the platform directly related to health, such as appointments, prescriptions, etc..