.
Updated the default SSL credentials to match local.ehealthi-doctors.com.

Updated to the latest version of the soul system, with better stability, and caching.

Updated to the latest version of modernuser, with new features.

Updated to the latest version of uniqueFileUpload, with better interoperability.

Updated default SSL credentials.

Corrected minor bug related to appointment notifications.

Improved typings.

Added prescription ability, both frontend, and backend.

Improved design of the app-patient-health widget.

Improved responsiveness of the device-frame widget.

Added ability for sub components of the device-frame widget, to move back and forth between views.

Improved appointment functionality by:
    Adding ability to chat during appointments.
    Adding notifications reminding doctors, and patients, of the appointments; as well as administrators, about new appointments.
    Adding ability to modify appointments, by an administrator.

Commenced functionality to have users pay for commodities.

Updated to the latest version of the Faculty of Chat, with stability, flexibility, and better UI.