.
Updated to the latest version of the soul system, with better networking, and stability.

Updated to the latest version of The Chat Faculty, with more robustness, and different approach to calling.

Improved prescription message sending algorithm.

Updated to the latest version of modernuser, with required features.

Improved prescription UI.

Maded general UI improvements.

Added quick-launch configuration for VS code.