.
Added ability to store health data. This is called medical records.

Updated to the latest version of the soul system, with better code quality.

Updated to the latest version of the faculty of chat, with a much needed feature to allow viewing of medical history from a chat.

Improved UX of security dashboard.