.
Improved code-completion.

Updated to the latest version of the soul system, with better code quality.

Created privacy-policy page, and added ability for admin to update it.

Updated to the latest version of modernuser, with bug fixes, and improvements.

Updated to the latest version of the faculty of chat, with better stability.