.
Updated to the latest version of modernuser, with better performance.

Updated to the latest version of chat, with bug fixes.

Improved design of home screen.

Added view for user notifications.

Updated to the latest version of the soul system, with bug fixes, better stability, and more code quality.

Added social media feeds feature.